Move Provenance to provenance.py and constraints to constraints.py
Extract Provenance class from core/distribution.py into
core/provenance.py (alongside existing graph utilities).

Extract Constraint classes, singletons, factories, and lattice logic
from core/distribution.py into new core/constraints.py.

Update all imports across the codebase to use canonical locations.

Closes #79
