+++ probpipe/__init__.py
@@ -34,7 +34,6 @@ from probpipe.distributions import (
     EmpiricalDistribution,
     ArrayEmpiricalDistribution,
     BroadcastDistribution,
-    Provenance,
     # Global settings
     BootstrapDistribution,
     JointBootstrapDistribution,
@@ -42,20 +41,6 @@ from probpipe.distributions import (
     RETURN_APPROX_DIST,
     set_default_num_evaluations,
     set_return_approx_dist,
-    # Constraints
-    Constraint,
-    real,
-    positive,
-    non_negative,
-    non_negative_integer,
-    boolean,
-    unit_interval,
-    simplex,
-    positive_definite,
-    sphere,
-    interval,
-    greater_than,
-    integer_interval,
     # Continuous
     Normal,
     Beta,
@@ -100,7 +85,22 @@ from probpipe.distributions import (
     LinearBasisFunction,
 )
 from probpipe.core.node import WorkflowFunction, Module, wf
-from probpipe.core.provenance import provenance_ancestors, provenance_dag
+from probpipe.core.provenance import Provenance, provenance_ancestors, provenance_dag
+from probpipe.core.constraints import (
+    Constraint,
+    real,
+    positive,
+    non_negative,
+    non_negative_integer,
+    boolean,
+    unit_interval,
+    simplex,
+    positive_definite,
+    sphere,
+    interval,
+    greater_than,
+    integer_interval,
+)
 from probpipe.modeling import Likelihood, GenerativeLikelihood, IterativeForecaster
 from probpipe.inference import (
     InferenceDiagnostics,
