+++ probpipe/converters/_scipy.py
@@ -13,9 +13,9 @@ import jax.numpy as jnp
 from ..core.distribution import (
     ArrayDistribution,
     ArrayEmpiricalDistribution,
-    Provenance,
     _auto_key,
 )
+from ..core.provenance import Provenance
 from ._protocol import Converter
 from ._registry import ConversionInfo, ConversionMethod
