+++ probpipe/core/provenance.py
@@ -1,14 +1,87 @@
-"""Provenance graph utilities for tracing distribution lineage."""
+"""Provenance tracking and graph utilities for distribution lineage."""
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from dataclasses import dataclass, field
+from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
     from .distribution import Distribution
 
-__all__ = ["provenance_ancestors", "provenance_dag"]
+__all__ = ["Provenance", "provenance_ancestors", "provenance_dag"]
 
+
+# ---------------------------------------------------------------------------
+# Provenance dataclass
+# ---------------------------------------------------------------------------
+
+@dataclass(frozen=True)
+class Provenance:
+    """Tracks how a distribution was created."""
+
+    operation: str
+    parents: tuple[Distribution, ...] = ()
+    metadata: dict[str, Any] = field(default_factory=dict)
+
+    def __repr__(self) -> str:
+        parent_names = ", ".join(
+            p.name or type(p).__name__ for p in self.parents
+        )
+        return f"Provenance({self.operation!r}, parents=[{parent_names}])"
+
+    # -- Serialization -----------------------------------------------------
+
+    def to_dict(self, *, recurse: bool = True) -> dict[str, Any]:
+        """Serialize to a JSON-compatible dict.
+
+        Parameters
+        ----------
+        recurse : bool
+            If True, recursively serialize parent provenance chains.
+            If False, only include parent type/name references.
+        """
+        parent_dicts = []
+        for p in self.parents:
+            entry: dict[str, Any] = {
+                "type": type(p).__name__,
+                "name": p.name,
+            }
+            if recurse and p.source is not None:
+                entry["source"] = p.source.to_dict(recurse=True)
+            parent_dicts.append(entry)
+
+        # Filter metadata to JSON-serializable values
+        safe_metadata = {}
+        for k, v in self.metadata.items():
+            if isinstance(v, (str, int, float, bool, list, dict, type(None))):
+                safe_metadata[k] = v
+            else:
+                safe_metadata[k] = str(v)
+
+        return {
+            "operation": self.operation,
+            "parents": parent_dicts,
+            "metadata": safe_metadata,
+        }
+
+    @classmethod
+    def from_dict(cls, d: dict[str, Any]) -> Provenance:
+        """Reconstruct from a dict produced by :meth:`to_dict`.
+
+        Parent distributions are not available at deserialization time, so
+        ``parents`` will be an empty tuple.  The parent information is
+        preserved in the dict under ``"parents"`` for inspection.
+        """
+        return cls(
+            operation=d["operation"],
+            parents=(),
+            metadata={**d.get("metadata", {}), "_parents_info": d.get("parents", [])},
+        )
+
+
+# ---------------------------------------------------------------------------
+# Graph utilities
+# ---------------------------------------------------------------------------
 
 def provenance_ancestors(dist: Distribution) -> list[Distribution]:
     """Return all ancestor distributions reachable via provenance chains.
